Enable edition 2018 (#57)
* Update MSRV in CI and Readme from 1.29 to 1.41.1

It seems we have consensus within the rust-bitcoin github organisation
on bumping the MSRV to 1.41.1, update CI job and the README to reflect
this.

* Update to use edition 2018

Add `edition = 2018` to the manifest file. No further changes were
necessary.

* Remove unneeded reference

Clippy emits:

  error: this expression borrows a reference (`&str`) that is
  immediately dereferenced by the compiler

As suggested, remove the unneeded reference.

* Use long literal separators

Clippy emits:

 warning: long literal lacking separators

As suggested, use separators.

* Use contains combinator instead of manual impl

Clippy emits:

 warning: manual `!RangeInclusive::contains` implementation

Use the code snippets suggested by clippy.

* Allow type_complexity

Clippy emits:

 warning: very complex type used. Consider factoring parts into `type` definitions

This code is in a unit test, its clear enough for the purpose.

Configure clippy to allow type_complexity

* Bump version 0.8.1 -> 0.9.0

We just bumped the MSRV, this is a major change but since we are pre 1.0
we just have to bump the minor version number.

Bump version from current `0.8.1` to `0.9.0`.