Add "alloc" feature
We would like users to be able to use parts of this library in a
`no_std` environment without an allocator. To achieve this add an
"alloc" feature and feature gate any code that requires allocation
behind "alloc"/"std".

Update the CI test job to run the test with each feature on its own.