Add gf32 module
We are attempting to get this crate to v0.1 (ie, API stabilization).

Currently we use a type call `u5` for the 5 bit math, while valid this
is slightly misleading because our `u5` is not basic integer type.
Within the bech32 encoding scheme the 5 bit type is an element of
GF32 (Galois Field).

Add a `gf32` module that defines a type `Fe32` (Field Element). This
type is similar to the current `u5` but includes add/sub/div/mul over
the finite field.

Introduce the `primitives` module to hold the `gf32` module (and other
future primitive types).

As yet do not make use of the `Fe32` type. Doing so is trivial and will
be done later after more primitive types are added.