use new `Fe32` type in place of `gf32`
This requires dropping `Ord` impls from the error types, but IMHO these
didn't really make sense anyway.

This also keeps the `u5` name. We will rename this in a later PR which
replaces the *Base32 traits with iterator adaptors, since at that point
there will be a much smaller API surface to change.