Remove unnecessary attribution
As we discussed in rust-bitcoin recently the "written by" attribute is
not that valuable because:

- It is not really true because many devs contribute to a file
- The 'by rust bitcoin devs' adds zero information

Remove the unnecessary attribution for all files in the `primitives`
module. Leave the `lib.rs` attribution as

    // Written by Clark Moody and the rust-bitcoin developers.

because Clark wrote the library initially and the rust-bitcoin devs came
along and hacked the s**t out of it, keep the attribution to give
recognition to Clark and to assume accountability by the rust-bitcoin
devs (well that's how I read it anyways).