Clean up comments in private error module
The `error` module is private so does not technically require module
level rustdoc, also the current rustdoc is too general (it is copy'pasta
from rust-bitcoin). Just remove it.

While we are at it, remove the 'Written by the Rust Bitcoin developers'
line because we have removed this everywhere else with the justification
that it is just noise.