Add hrpstring and iter modules
Currently there are a bunch of types and traits in `lib.rs` for
converting to/from base32 (incl. doing so without allocating).

Observe that instead of custom types we can use iterator adapters to
achieve this. Some benefits are:

- More composabality
- More "rusty"
- Uses the new `checksum` module (i.e, better separation of concerns)
- Hopefully more maintainable

Add two new sub-modules to the `primitives` module:

- `iter`: Implement various adaptor iterators for allocationless
  checksumming etc.
- `hrpstring`: Add the `Parsed` struct that can be used for parsing
  bech32 address strings.

Remove everything that is now not needed from `lib.rs` while making an
effort not to change the API too much i.e., keep the current public
functions more or less as they are, with the following notable
exceptions:

- `encode_to_fmt` no longer returns a result
- `encode` now returns a `fmt::Error` as the `Result` error
  variant (instead of `Error`).
- `decode` returns a `Parsed` instead of an HRP and vector of u5s, to get the
  HRP/u5's one can use the `hrp()` and `fe32_iter()` functions on `Parsed`.
- `decode_without_checksum` return value changed as per `decode`.
- The crate level `Error` type is totally re-written.

All the recently added but as-yet-unreleased `arrayvec` stuff which was
added to improve alloc free encoding/decoding is removed.

Complete list of the public types/functions removed in this patch - (this
functionality is now provided by `primitives`):

- traits:
  - `WriteBase32`
  - `WriteBase256` (unreleased)
  - `FromBase32`
  - `ToBase32`
  - `Base32Len`
  - `CheckBase32`

- types (and their trait implementations):
  - `Bech32Writer`
  - `ComboError` (unreleased)
  - `Case` (exist still but is private)

- functions:
  - `convert_bits`
  - `convert_bits_in` (unreleased)
  - `encode_to_fmt_anycase` (unreleased)
  - `decode_lowercase` (unreleased)