decode: Add empty data check
Currently we index into the data field without first checking it is not
empty. For context, the indexing is done _before_ we do segwit validity
checks which check for correct data lengths.

This is a real noob bug , bad Tobin - no biscuit.

Add a check for the empty array before indexing into it.