Rename EncodeError::Write variant to Fmt
We typically refer to a "fmt" being an object that implements
`fmt::Write` e.g, `encode_to_fmt()`.

Currently we are using the variant name `Write` for an error variant
that holds a `fmt::Error`, while this is correct in a way it is also
confusing because we may one day have an `io::Error` (coming from a
`io::Write` writer).

Use `EncodeError::Fmt` instead of `Write`.