Enforce maximum string length
BIP-173 states that a bech32 string must not exceed 90 characters but
this is a bip specification, the bech32/bech32m checksum algos can
actually cover more characters than that.

To keep the `segwit` stuff bip compliant and also keep the library
general add an associated `CODE_LENGTH` const to the `Checksum` trait.
Then in the `segwit` module and types enforce the 90 character limit but
in the general API and modules enforce the `Ck::CODE_LENGTH` limit.

FTR in `bech32 v0.9.0` no lengths were not enforced.