segwit: Add a buffer when writing
Currently we call `fmt::Write::write_char` and `io::Write::write_all`
in a loop while iterating over the encoded characters of a bech32
string. This is inefficient.

Add a buffer and copy ASCII bytes into it while looping the write the
whole buffer out in a single call. The buffer has a size of 90 bytes
because we know the bech32 string is guaranteed to be 90 chars or less.