Use pass-by-copy for Hrp when possible
We currently have some code that uses pass-by-reference and some that
uses pass-by-copy.

Further investigation shows that when messing with iterators we have to
have a reference to the `Hrp` so that the iterator has somewhere in
memory to point to while iterating.

The iterator API is intentionally low level, we can make the higher
level APIs more ergonomic by using pass-by-copy.