primitives: introduce `Polynomial` type, generator polynomial analysis
This introduces a type `Polynomial` which wraps a vector of field
elements which it treats as the coefficients of a polynomial. It exposes
only an ad-hoc set of functionality needed for error correction and is
not intended to be a general-purpose polynomial type, so it is not
exposed in the public API.

Because it is backed by a `Vec` the type and its module are only
available when the alloc feature is available.