primitives: introduce Field trait
This new trait covers a larger matrix of ops, and implements them all
with a single macro. Will allow us to introduce new fields (in
particular, extension fields of GF32) in a future commit with more
consistent/less boilerplate heavy implementations of basic arithmetic.

We run the API checker in this commit so you can see that the API
changes are strictly additive and only add missing op implementations.