primitives: introduce FieldVec type
This will allow `Polynomial` to work without an allocator (at least for
"small" checksums, including every checksum supported by this library).
Here a "small" checksum is one of length at most 6 (which covers bech32
and bech32m).

codex32 (13 characters) and "long codex32" (15 characters) will not work
with no-alloc. I would kinda like to fix this but it results in large
types, especially a large InvalidResidueError type, so probably it will
need to be feature-gated or something. For now we just punt on it.