primitives: introduce FieldVec type
This will allow `Polynomial` to work without an allocator (at least for
"small" checksums. Here a "small" checksum is one of length at most 6
(which covers bech32 and bech32m).

The descriptor checksum (8 characters), codex32 (13 characters) and "long
codex32" (15 characters) will not work with no-alloc.

I would like to fix this but it results in large types, especially a large
InvalidResidueError type (see last commit of this PR), so probably it will
need to be feature-gated or something. For now we just punt on it.

This PR introduces the `primitives::correction` module but only puts a
single constant into it for now.