primitives: introduce the Berlekamp-Massey algorithm for computing linear shift registers
This provides a general-purpose implementation of the Berlekamp-Massey
algorithm for finding a linear shift register that generates a given
sequence prefix.

If compiled without an allocator, it will run less efficiently (and be
limited to a maximum size) but it will work.

Also introduces a fuzz test to check that it works properly and does not
crash.