field: add Default bound to Field
In the next commits we'll introduce a generic no-alloc container called
FieldVec whose implementation will be much nicer if we have a Default
bound on all of our field elements. This way we can implement the
"FieldVec", which despite the name really has nothing to do with fields,
purely in terms of core traits, rather than confusing readers of the
code by having Field bounds everywhere without ever doing algebra.

Since fields all have a ZERO member, which is a sensible output for
Default, this bound is reasonable to require.