field: split trait into "normal" field trait and sealed trait
Several methods (and we are going to add more) don't really belong in a
general-purpose field trait, which really is just "a type that can be
multiplied and added and has some associated constants".

This isn't a general-purpose math library, but fields are a pretty
common thing for people to want. And it's pretty easy for us to expose
this trait in a nice way, so we might as well do it.