field: use UFCS in macro
This is purely a code refactor to get rid of a `use` statement (which
will trigger an "unused code" warning if the macro is ever called
somewhere where the Field trait is already in scope). As a side effect
it reduces the size of the macro by roughly 75% in terms of line count.