polynomial: make Eq/PartialEq take leading zeros into account
Two polynomials are equal even if they have a different number of
leading zeros.

Putting this into its own commit rather than folding it into one of the
FieldVec commits, because this bug was present even before the FieldVec
stuff.

Also remove the `Hash` impl because it's unneeded and we would be
expected to keep it in sync with Eq.